feat(swc-plugin-gql): remaining changes for plugin
