+++ packages/swc-plugin-gql/.gitignore
