+++ packages/swc-plugin-gql/.gql-query-whitelist
@@ -1,2 +1,3 @@
 dbdf391e40701a11ce7153fbfce1e86f
 85de686dd7c8c76de42f53a85e9e47b1
+1b1b09de6ddff26791eca97213136590
