+++ packages/swc-plugin-gql/fixtures/fragments/ImageUrls.gql
@@ -1,0 +1,4 @@
+fragment ImageUrls on Image {
+  baseUrl
+  secureBaseUrl
+}
