+++ packages/swc-plugin-gql/fixtures/query.gql
