+++ packages/swc-plugin-gql/tests/__swc_snapshots__/src/lib.rs/test_transform_gql_import.js
@@ -1,0 +1,2 @@
+const GET_CONFIG = `{ organization(login: "facebook") { name } }`;
+export default GET_CONFIG;
