+++ packages/swc-plugin-gql/tests/__swc_snapshots__/src/lib.rs/test_transform_gql_import_with_sub_imports.js
@@ -1,0 +1,25 @@
+const GET_CONFIG = `
+query GetConfig {
+  config {
+    ...ConfigImages
+  }
+}
+
+
+fragment ConfigImages on Config {
+  images {
+    backdropSizes
+    logoSizes
+    posterSizes
+    profileSizes
+    stillSizes
+    ...ImageUrls
+  }
+}
+
+fragment ImageUrls on Image {
+  baseUrl
+  secureBaseUrl
+}
+`;
+export default GET_CONFIG;
