+++ packages/swc-plugin-gql/types.d.ts
@@ -1,0 +1,4 @@
+type ImportGql = (path: TemplateStringsArray) => string;
+
+export = ImportGql;
+export default ImportGql;
